+++ shepherd_core/shepherd_core/data_models/experiment/observer_features.py
@@ -11,19 +11,16 @@ from pydantic import PositiveFloat
 from pydantic import model_validator
 from typing_extensions import Self
 
-from shepherd_core import logger
 from shepherd_core.data_models.base.shepherd import ShpModel
 from shepherd_core.data_models.testbed.gpio import GPIO
+from shepherd_core.logger import log
 
 # defaults (pre-init complex types)
 zero_duration = timedelta(seconds=0)
 
 
 class PowerTracing(ShpModel, title="Config for Power-Tracing"):
-    """Configuration for recording the Power-Consumption of the Target Nodes.
-
-    TODO: postprocessing not implemented ATM
-    """
+    """Configuration for recording the Power-Consumption of the Target Nodes."""
 
     intermediate_voltage: bool = False
     """
@@ -38,15 +35,15 @@ class PowerTracing(ShpModel, title="Config for Power-Tracing"):
 
     default is None, recording till EOF"""
 
-    # post-processing
-    calculate_power: bool = False
-    """ ⤷ reduce file-size by calculating power -> not implemented ATM"""
+    # further processing of IV-Samples
+    only_power: bool = False
+    """ ⤷ reduce file-size by calculating power and automatically discard I&V"""
     samplerate: Annotated[int, Field(ge=10, le=100_000)] = 100_000
-    """ ⤷ reduce file-size by down-sampling -> not implemented ATM"""
-    discard_current: bool = False
-    """ ⤷ reduce file-size by omitting current -> not implemented ATM"""
-    discard_voltage: bool = False
-    """ ⤷ reduce file-size by omitting voltage -> not implemented ATM"""
+    """ ⤷ reduce file-size by re-sampling (mean over x samples)
+        Timestamps will be taken from the start of that sample-window.
+        example: 10 Hz samplerate will be binning 10k samples via mean() and
+                 the timestamp for that new sample will be value[0] of the 10k long vector
+    """
 
     @model_validator(mode="after")
     def post_validation(self) -> Self:
@@ -55,22 +52,10 @@ class PowerTracing(ShpModel, title="Config for Power-Tracing"):
         if self.duration and self.duration.total_seconds() < 0:
             raise ValueError("Duration can't be negative.")
 
-        discard_all = self.discard_current and self.discard_voltage
-        if not self.calculate_power and discard_all:
-            raise ValueError("Error in config -> tracing enabled, but output gets discarded")
-        if self.calculate_power:
-            raise NotImplementedError(
-                "Feature PowerTracing.calculate_power reserved for future use."
-            )
-        if self.samplerate != 100_000:
-            raise NotImplementedError("Feature PowerTracing.samplerate reserved for future use.")
-        if self.discard_current:
-            raise NotImplementedError(
-                "Feature PowerTracing.discard_current reserved for future use."
-            )
-        if self.discard_voltage:
-            raise NotImplementedError(
-                "Feature PowerTracing.discard_voltage reserved for future use."
+        rates_allowed = (10, 100, 1_000, 100_000)
+        if self.samplerate not in rates_allowed:
+            raise ValueError(
+                "Feature PowerTracing.samplerate only supports specific rates: %s", rates_allowed
             )
         return self
 
@@ -188,7 +173,7 @@ class GpioTracing(ShpModel, title="Config for GPIO-Tracing"):
         if self.duration and self.duration.total_seconds() < 0:
             raise ValueError("Duration can't be negative.")
         if self.uart_decode:
-            logger.error(
+            log.error(
                 "Feature GpioTracing.uart_decode reserved for future use. "
                 "Use UartLogging or manually decode serial with the provided waveform decoder."
             )