+++ shepherd_core/shepherd_core/data_models/task/observer_tasks.py
@@ -1,11 +1,15 @@
 """Collection of tasks for selected observer included in experiment."""
 
+from collections.abc import Set as AbstractSet
 from datetime import datetime
-from datetime import timedelta
 from pathlib import Path
+from pathlib import PurePosixPath
+from typing import Annotated
+from typing import Optional
 
 from pydantic import validate_call
 from typing_extensions import Self
+from typing_extensions import deprecated
 
 from shepherd_core.data_models.base.content import IdInt
 from shepherd_core.data_models.base.content import NameStr
@@ -15,6 +19,7 @@ from shepherd_core.data_models.testbed.testbed import Testbed
 
 from .emulation import EmulationTask
 from .firmware_mod import FirmwareModTask
+from .helper_paths import path_posix
 from .programming import ProgrammingTask
 
 
@@ -24,17 +29,21 @@ class ObserverTasks(ShpModel):
     observer: NameStr
 
     # PRE PROCESS
-    time_prep: datetime  # TODO: should be optional
+    time_prep: Optional[datetime] = None  # TODO: currently not used
     root_path: Path
 
     # fw mod, store as hex-file and program
-    fw1_mod: FirmwareModTask | None = None
-    fw2_mod: FirmwareModTask | None = None
-    fw1_prog: ProgrammingTask | None = None
-    fw2_prog: ProgrammingTask | None = None
+    fw1_mod: Optional[FirmwareModTask] = None
+    fw2_mod: Optional[FirmwareModTask] = None
+    fw1_prog: Optional[ProgrammingTask] = None
+    fw2_prog: Optional[ProgrammingTask] = None
 
     # MAIN PROCESS
-    emulation: EmulationTask | None = None
+    emulation: Optional[EmulationTask] = None
+
+    # deprecations, TODO: remove before public release
+    owner_id: Annotated[Optional[IdInt], deprecated("not needed anymore")] = None
+    abort_on_error: Annotated[bool, deprecated("has no effect")] = False
 
     # post_copy / cleanup, Todo: could also just intake emuTask
     #  - delete firmwares
@@ -48,24 +57,14 @@ class ObserverTasks(ShpModel):
         if not tb.shared_storage:
             raise ValueError("Implementation currently relies on shared storage!")
 
-        t_start = (
-            xp.time_start
-            if isinstance(xp.time_start, datetime)
-            else datetime.now().astimezone() + timedelta(minutes=3)
-        )  # TODO: this is messed up, just a hotfix
-
         obs = tb.get_observer(tgt_id)
-        xp_dir = "experiments/" + xp.name + "_" + t_start.strftime("%Y-%m-%d_%H-%M-%S")
-        root_path = tb.data_on_observer / xp_dir
-        # TODO: Paths should be "friendlier"
-        #    - replace whitespace with "_" and remove non-alphanum?
-
+        root_path = tb.data_on_observer / "experiments" / xp.folder_name()
         fw_paths = [root_path / f"fw{_i}_{obs.name}.hex" for _i in [1, 2]]
 
         return cls(
             observer=obs.name,
-            time_prep=t_start - tb.prep_duration,
-            root_path=root_path,
+            # time_prep=
+            root_path=path_posix(root_path),
             fw1_mod=FirmwareModTask.from_xp(xp, tb, tgt_id, 1, fw_paths[0]),
             fw2_mod=FirmwareModTask.from_xp(xp, tb, tgt_id, 2, fw_paths[1]),
             fw1_prog=ProgrammingTask.from_xp(xp, tb, tgt_id, 1, fw_paths[0]),
@@ -93,3 +92,12 @@ class ObserverTasks(ShpModel):
                 raise ValueError("Emu-Task should have a valid output-path")
             values[self.observer] = self.emulation.output_path
         return values
+
+    def is_contained(self, paths: AbstractSet[PurePosixPath]) -> bool:
+        all_ok = any(self.root_path.is_relative_to(path) for path in paths)
+        all_ok &= self.fw1_mod.is_contained(paths)
+        all_ok &= self.fw2_mod.is_contained(paths)
+        all_ok &= self.fw1_prog.is_contained(paths)
+        all_ok &= self.fw2_prog.is_contained(paths)
+        all_ok &= self.emulation.is_contained(paths)
+        return all_ok